integer timestamps + remove old cached invalid transactions

+++ tcoin/p2p/requests/request.py
@@ -13,7 +13,7 @@ class Request(SignedPayload):
         *,
         node_id: str,
         payload: dict,
-        timestamp: float = None,
+        timestamp: int = None,
         hash: str = None,
         signature: str = None,
         response: dict = None,
